+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.0.0rc5 - 2024-03-03
+### Fix
+* logging for Simulate mode
+
+### Update
+* Binance: handling `EXPIRED_IN_MATCH`
+
 ## 3.0.0rc1 - 2024-03-02
 ### Fix
 # TODO * `cancel_order_call()`: not raised asyncio.TimeoutError on timeout